Dashboard: order items details in progress.

+++ resources/views/dashboard/product/index.blade.php
@@ -1,34 +1,54 @@
-<x-app-layout>
+@php
+    $breadcrumbs = [
+        [
+            'name' => 'Dashboard',
+            'link' => route('dashboard'),
+        ],
+        [
+            'name' => 'Products',
+            'link' => false,
+        ],
+        [
+            'name' => 'Listing',
+            'link' => false,
+        ],
+    ];
+    $title = 'Products';
+@endphp
+
+@extends('layouts.app', [
+    'title' => $title, 
+    'breadcrumbs' => $breadcrumbs,
+    'addButton' => true,
+    'btn' => [
+        'text' => 'Add Product',
+        'link' => route('product.create')
+    ]
+])
+
+@section('title', $title)
+
+@push('styles')
     <style>
         .delete,
         .edit {
             height: 40px;
             width: 40px;
             border-radius: 50%;
-            color: white
-        }
-
-        .delete:hover,
-        .edit:hover {
-            color: white
         }
 
         .image {
             width: 50px;
             height: 50px;
         }
+
+        .swal2-popup .swal2-styled {
+            margin: 0 10px !important;
+        }
     </style>
-    
-    <x-slot name="header">
-        <div class="d-flex justify-content-between align-items-center">
-            <h2 class="font-semibold text-xl text-gray-800 leading-tight">
-                {{ __('Products') }}
-            </h2>
+@endpush
 
-            <a href="{{ route('product.create') }}" class="btn btn-primary">Add Product</a>
-        </div>
-    </x-slot>
-
+@section('content')
     <div class="container-fluid">
         <div class="p-5">
             <div class="table-responsive">
@@ -59,11 +79,11 @@
                                             <a class="delete bg-danger d-flex align-items-center justify-content-center"
                                                 href="javascript:void(0)"
                                                 data-url="{{ route('product.destroy', $product->id) }}">
-                                                <i class="fas fa-trash-alt"></i>
+                                                <i class="fas fa-trash-alt text-white"></i>
                                             </a>
-                                            <a class="edit ml-2 bg-primary d-flex align-items-center justify-content-center"
+                                            <a class="edit ms-2 bg-primary d-flex align-items-center justify-content-center"
                                                 href="{{ route('product.edit', $product->id) }}">
-                                                <i class="fas fa-edit"></i>
+                                                <i class="fas fa-edit text-white"></i>
                                             </a>
                                         </div>
                                     </td>
@@ -71,7 +91,7 @@
                             @endforeach
                         @else
                             <tr>
-                                <td colspan="7" class="text-center">
+                                <td colspan="5" class="text-center">
                                     <b>
                                         No products created
                                     </b>
@@ -83,32 +103,42 @@
             </div>
         </div>
     </div>
+@endsection
 
-    @section('scripts')
-        <script>
-            $('.delete').on('click', function() {
-                axios.delete($(this).data('url'))
-                    .then(response => response.data)
-                    .then(data => {
-                        if (data.success) {
-                            window.location.reload()
-                        } else {
+@push('scripts')
+    <script>
+        $('.delete').on('click', function() {
+            swal.fire({
+                icon: 'warning',
+                title: 'Confirm delete',
+                text: 'Are you sure to proceed?',
+                showCancelButton: true,
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    axios.delete($(this).data('url'))
+                        .then(response => response.data)
+                        .then(data => {
+                            if (data.success) {
+                                window.location.reload()
+                            } else {
+                                swal.fire({
+                                    icon: 'error',
+                                    title: 'Oops...',
+                                    text: 'Something went wrong!',
+                                })
+                            }
+                        })
+                        .catch(err => {
                             swal.fire({
                                 icon: 'error',
                                 title: 'Oops...',
                                 text: 'Something went wrong!',
                             })
-                        }
-                    })
-                    .catch(err => {
-                        swal.fire({
-                            icon: 'error',
-                            title: 'Oops...',
-                            text: 'Something went wrong!',
                         })
-                    })
+                }
             })
-        </script>
-    @endsection
 
-</x-app-layout>
+
+        })
+    </script>
+@endpush
